Support Lantronix EDS3000-series console server
Our Lantronix EDS3000-series serial console server creates a TCP port
per serial line to connect directly to that serial line.

Teach the sonic-mgmt infrastructure about this type of console.

Notably, there are two aspects of console server sonic-mgmt needs to
know about. The first is the serial-line aspect, where we ssh directly
to the TCP port which corresponds to the serial line. The second is
the "menu_config" aspect which is used to forcefully terminate
previous users of the ssh-serial connection.