ISS-940 Failing test: stress/test_stress_routes.py
Rootcause:
announce_withdraw_routes function stores all memory and other statistics
in a folder and use it for analyzing the statistics before and after
stress configurations. This function uses explicit folder called 'logs'
to store these statistics, but this folder is not available in the system
which is causing error and further analysis failure.

Resolution:
Either we need to make sure that the 'logs' folder is created in
current module ('tests') folder otherwise, leave this argument
so that the anncounce_routes function will use default folder
which is available in the system. Making the change consistent
to other places where the announce_routes is used without
passing any explicit log_path value.

wq