ISS-701: Enable telemetry in test code (#71)
Move the code to enable telemetry if it's not running to a codepath that
all telemetry tests will use. Note that there is an actual testcase in
test_telemetry.py that checks if telemetry is enabled by default, so we
probably need to fix this in our infrastructure setup and not just in
the tests. But this works for now, and gets all the telemetry tests up
and running.