ISS-653: use makedirs() instead of mkdir() to make log dirs (#72)
Currently, the code uses os.mkdir() to build the directory structure
needed to house the files generated by telemetry/test_events.py. mkdir()
fails if parts of the path do not exist. The base directory for the
paths used in this test is 'logs'. This directory does not exist by
default, and the code assumes that it does (maybe because other tests
might have created it?). If we use makedirs() instead, we do not hit
this issue.