ISS-1250 tests/bgp/test_bgp_peer_shutdown.py failure
Error #1 is due to the timestamp comparison between test server vs
DUT which are running in two different time zones.
The timestamp_before_teardown is taken from test server and
bgp_session_down_time is taken from DUT's log. The test server is
running in PDT and DUT is running in UTC.

Error #2 is due to log analysis failure due to static-route which
will be addressed in a separate PR.

Resolution: timestamp_before_teardown is updated with DUT's timestamp.