ISS-828 sonic-mgmt Enable bgp frr unified mode conditionally in deploy-mg
This is follow-up PR of commit ID:  6faa928f0daceb96edff9c0268426d66b5312713

With early commit, the deploy-mg was failing for virtual testbed.
Hence, we commented out the calling part of the script.  Now, the code is
updated to handle the failure conditions of virtual testbed as well.

For detail of this feature we can check the PR description of above
mentioned commit ID.