ISS-1561 [frr mgmt framework mode] test_bgp_fact.py failure (#188)
Why I did it
When FRR management framework is enabled, the bgp fact
schema storing the neighbor detail will be different.

How I did it
Updated the code to handle the schema change when
FRR management framework is enabled.

How to verify it
Enable FRR management framework config mode in the
DUT and run the test_bgp_fact.py test.