NO-ISS: Skip test_ser.py on TH5 Nexthop platforms (#207)
* NO-ISS: Skip test_ser.py on TH5 Nexthop platforms

The functionality test_ser.py is testing is not supported on TH5. Therefore, we should skip it.

There's an upstream bug that was tracking this for skipping this test on Arista TH5 platforms here: https://github.com/sonic-net/sonic-mgmt/issues/17459

However, that bug was closed (it looks like it shouldn't have been yet), so I'm not conditioning this SKIP on that bug.

* review feedback

use "hwsku in broadcom_th5_hwskus" instead.