ISS-1608: [gcu] Skip the ECN test when the DUT doesnt support the config (#269)
What is the motivation for this PR?
When the ECN configuration is not supported in the base configuration
generated for a platform, the test should be get skipped.

The base configuration is generated based on the qos.json.j2 (refered from)
/usr/share/sonic/device/<Platform>/<HwSKU>/qos.json.j2.
When there is no matching platform template exists,
 the configuration doesnt get generated and without this the test fails.

How did you do it?
Skipped the test when not applicable

How did you verify/test it?
Verified the test gets skipped in the DUTs without the ECN configuration.