ISS-1918: Configure the NTP in mgmt VRF before verification (#270)
What is the motivation for this PR?
The ntp test fails as the service configuration has to be reinitialized to run with management VRF.

How did you do it?
While its possible to just run the chrony or the ntp directly with ip vrf exec mgmt (aka execute_dut_command()). Doing the configuration through json also validates the service file generation and tests the same. Hence the change to patch the configuration to use management VRF and reset the same once the test is done.

How did you verify/test it?
Verified the test_ntp passes in a prepared DUT.