NO-ISS: Make apt-get install non-interactive in script (#294)
What is the motivation for this PR?
In a virtual testbed, when setting up sonic-mgmt, the script attempts to install the bridge-utils, which requires some services to be restart and since the install is invoked by a script, the TUI waits for an input and doesnt proceed.

How did you do it?
Use noninteractive mode and also set NEEDRESTART_MODE to automatically restart services when required.

How did you verify/test it?
Verified the prepare testbed is successful