ISS-1919: [autorestart] Reset start-failed for swss for syncd test (#315)
What is the motivation for this PR?
The tests/autorestart/test_container_autorestart.py fails

How did you do it?
When the critical process in the syncd is killed and container gets stopped, the autorestart doesnt happen since swss service hits the start limit and without swss starting, the syncd doesnt get started.

The test already handles the case where start-limit is hit for the service that is being checked for. But syncd is a special case where the swss service state needs to be checked instead of syncd.

How did you verify/test it?
Verified the test passes with the change