ISS-716: Make GNMI Config DB tests VRF aware  (#329)
Parameterizing `test_gnmi_configdb.py` to be able to work in different VRF scenarios.

Note:
Since `eth0` is the `MGMT_INTERFACE`, when `mgmt` VRF is enabled the interface gets bound to the VRF automatically. Hence, for that case the gNMI client needs to run in the VRF context. But, the `gnmi` docker container doesn't have elevated privileges (it doesn't get started with `--privileged` option), and so we have to run the client from the host.