ISS-2353: Parameterize only GNMI configdb tests with VRF (#365)
GNMI tests were failing with mvrf enabled after the tests were parameterized to use different VRF configs. This is because enabling mvrf has other feature interactions, and therefore all test setup & validation logic needs to be updated to handle mvrf scenarios, which was beyond the scope of the original ISS-716 ticket where the tests got parameterized.

This change only keeps `test_gnmi_configdb.py` parameterized with different VRF configs, as that test is relevant to the VRF aware GNMI changes which was the original goal of ISS-716.