ISS-2398: [radius] Include Message-Authenticator in response from server (#374)
Summary:
Fixes # ISS-2398

Type of change
 Bug fix

What is the motivation for this PR?
With the change in upstream in SONiC to include Message-Authenticator, it also expects by default that the server would send the message-authenticator attribute in responses. Without that the client ignores the received response causing the failures.

How did you do it?
The Message-Authenticator attribute inclusion can be done in post-auth process in radius server controlled using the configuraiton in sites-enabled/default. This configuration file is now overridden with additional configuration in post-auth to check the request and include the Message-Authenticator if necessary.

The sections with the below code are addon to the original default configuration.
```
       if (&request:Message-Authenticator) {
               update reply {
                       Message-Authenticator := 0x00
               }
       }
```
How did you verify/test it?
```
[INFRA] 2025-07-31 23:01:37: run_all_tests():
TESTS RUN RESULTS:BEG
+----------------------------------+-----------------------------------+--------+----------+-------+
| Test                             | Description                       | Result | Duration | Error |
+----------------------------------+-----------------------------------+--------+----------+-------+
| tests/radius/test_radius_auth.py | skipped 0 errors/failed 0 total 6 | PASS   | 421.08   |       |
+----------------------------------+-----------------------------------+--------+----------+-------+
TESTS RUN RESULTS:END
```