ISS-2832 Move all expected regexes to ignored to support multiple sairedis modes (#405)
### Description of PR
The previous attempt
https://github.com/nexthop-ai/private-sonic-mgmt/pull/387 worked fine
for the updated sonic-sairedis mode, but not the current code in
sonic-sairedis. This change makes all regexes as ignored to make the
test pass. Once the sonic-sairedis pointer is updated, we can then move
the correct ones to the expected regex.

Summary:
Fixes # (issue)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

### Approach
#### What is the motivation for this PR?

#### How did you do it?

#### How did you verify/test it?

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->