ISS-2960: Add workaround for ACL limit issue (#413)
Summary:
Some tests which use ACLs can fail when all ACLs have been programmed.
We can detect this case and compensate for it by storing the table info
and re-instating it at the end of the test.

Fixes 2960

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?

Increase code coverage.

#### How did you do it?

In the test infra, store the contents of the DATAACL table and restore
those contents at the end of the test.

#### How did you verify/test it?

Test fails without this change,  but passes with the change.

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->