ISS-3070: Fix test_bgp_sentinel.py failure (#462)
### Description of PR
The issue was that BGPMON and ipv6 are not supported and the test
reached that conclusion rather slowly. In the meantime it does configure
a whole bunch of things clobbering other tests. This PR helps the test
reach the conclusion to skip the tests faster. The nice side effect is
that the test runs a little bit faster.


Summary:
Fixes # ISS-3070

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?
Failure of test_bgp_sentinel.py

#### How did you do it?
Skipped certain parameters for the test earlier

#### How did you verify/test it?
By running it against the dut

#### Any platform specific information?
None

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->