ISS-3033: Fix v6 mirror type in everflow tests (#475)
### Description of PR
Everflow test is is not setting mirror type to ERSPAN for V6 mirror
sessions.
This is causing v6 mirror sessions to be programmed in-correctly in
ASIC.

Summary:
Fixes # (issue)
Set the mirror type to ERSPAN while creating

### Type of change


- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

### Approach
#### What is the motivation for this PR?

#### How did you do it?

#### How did you verify/test it?
Everflow v6 tests passing


#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->