ISS-835-1: Add sonic mgmt remote attestation test (#488)
### Description of PR

Summary: Adding a new test case to run TPM Remote Attestation. 

The test needs specific packages in the `docker-sonic-mgmt` container in
order to work:
- Recent enough `tpm2-tools` >= 5.x
- `asn1crypto==1.5.1`
- `cryptography>=46.0.1`
- click

Either we merge
https://github.com/nexthop-ai/private-sonic-buildimage/pull/1759 and
start building the container in our CI or we must run the environment
setups as part of the test case. This is dependent on the container
instance itself being cleaned up so we don't leave a dirty container for
later testing.

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [x] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

### Approach
#### What is the motivation for this PR?

Adding a new test case to run TPM Remote Attestation, so we can
continuously validate the attestation workflow

#### How did you do it?

Added a new `sonic-mgmt` test to:
- Generate the PCR sets using the `golden_pcr_generation.py`
- Run `tpm_remote_attestation.py` with the generated PCR sets

#### How did you verify/test it?

[test_tpm_remote_attestation.log](https://github.com/user-attachments/files/22545473/test_tpm_remote_attestation.log)

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?
Test simply runs on any TPM enrolled Nexthop DUT.

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->