ISS-2067: Skip mirror policer handling for Nexthop devices (#546)
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit
easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should
reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
We don't support mirror policer feature on Nexthop devices yet. There is
existing machinery in this test to allow the mirror policer codepath to
be skipped if it's not supported. We should add Nexthop devices to the
list of skipped devices. We will have to remember to come back and
remove it when we add support for mirror policer. We don't want to
ignore the failures, and we don't want to silence them in SAI either if
we don't have to, so modifying the test code is the best approach.

Fixes # (issue)

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [x] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?

This test fails due to SAI error logs about feature availability, and we
lose test coverage because of that.

#### How did you do it?

Skip the policer code. There is already an existing mechanism to do
that.

#### How did you verify/test it?

The test passes with these changes.

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->