ISS-2067: Revert "ISS-2067 : Added logRe to ignore mirror session policer errors " (#547)
Reverts nexthop-ai/private-sonic-mgmt#520

@bhouse-nexthop and @avinash-nexthop brought up a good point, that if we
silence these loganalyzer failures we won't be able to claim this test
on the passrate for customers, since if they were to run the test
themselves without the ignore logre they would see the failure. So we
need to find another mechanism to address this, since we aren't going to
add support for this feature any time soon.

It was suggested that we could handle this at the SAI level, but I think
that's just as bad as adding the ignore, since we are still swallowing
errors in that case but with even less visibility into why or how that's
happening. Instead, we should solve this at the test level, and just
skip the policer-specific codepath for Nexthop devices until the feature
is supported. This should be acceptable upstream. The only downside is
that we will have to remember to come back and fix the test up when the
feature finally does get supported.

See [this
change](https://github.com/nexthop-ai/private-sonic-mgmt/pull/546) for
the test fix approach.

This needs to be backported to 202411 and 202505 as well.