ISS-3620: V6 mirror tests need to be skipped for per interface tests (#552)
### Description of PR
test_everflow_per_interface.py should be skipped for V6 mirror as
mentioned in
https://github.com/sonic-net/sonic-swss/issues/2204. This bug was closed
recently which seems to be causing the tests to be run now. These can
not be enabled though since this fix is still missing in master:
https://github.com/sonic-net/sonic-swss/pull/2668/files.

Summary:
Fixes # (issue)
Explicitly skipping the test irrespective of the
https://github.com/sonic-net/sonic-swss/issues/2204 status.

### Type of change


- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

### Approach
#### What is the motivation for this PR?
test_everflow_per_interface v6 tests are failing.

#### How did you do it?
Fixed the skip condition for the tests to skip for v6.

#### How did you verify/test it?
Ran and verified the tests are skipped for v6

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->