NOS-135: fix intermittent failures when looking for metadata (#580)
### Description of PR
The packet watcher skips packets on src port, when trying to verify
metadata insertions. But on TH5, the src port and egress port can be the
same, as the INT packet is considered a new packet with added bytes and
it can hash to the same ingress port. This caused intermittent test
failures.

### Type of change
- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
Remove the packet reflection check and increase the timeout to 5 seconds

#### What is the motivation for this PR?
Fix intermittent test failures on tam/test_ifa.py

#### How did you verify/test it?
Ran the test in a loop to make sure it passes consistently.

#### Any platform specific information?
Test is applicable only for TH5