ISS-2777:Fixing the test to use correct IP addresses (#605)
### Description of PR
There is a BCM SAI bug where the source and dest IP address in the MOD
packets are in host byte order, instead of network byte order. In this
test I was using IP addresses 1.1.1.1 and 2.2.2.2 which masked the
problem. Here I am changing the IP addresses to not mask the issue.

Summary:
Fixes # (issue)
1. Using IP addresses that would fail if the order was not matched.
2. While debugging this issue, I also had to fix other parts of the test
where the generated test had errors which were masked

### Type of change
- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?

#### How did you do it?

#### How did you verify/test it?

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->