NOS-136: [mvrf] Reconfigure snmp agent to listen on mgmt vrf (#608)
### Description of PR
Change the `snmpagentaddress` to bind to `mgmt` VRF when `mgmt` VRF is
created.

Summary:
Fixes # NOS-136

### Type of change
- [x] Bug fix

### Back port request
- [x] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?
The tests involving the reboot in `test_mvrf.py` were failing as the
`snmpagentaddress` configuration were dictating the bind on `default`
VRF eventhough the interface is moved to `mgmt` VRF.

#### How did you do it?
SNMP agent address bind configuration is separate from mgmt VRF
configuration. So, when the mgmt VRF is configured, accordingly the SNMP
binding configuration also has to be changed to reflect the same.
Without that the snmp bind fails and would prevent coming up.

This ends up in supervisord to give up and exit, causing the docker
container to go down which inturn makes systemd to retry till it enters
failed state.

The proposed change, configures the snmpagentaddress to loose the
previous management address bind (on the default VRF) and instead bind
on the mgmt VRF when the VRF gets configured.

Note that, when snmpagentaddress entry is added/removed, it triggers
snmp service to be restarted, which means, when 2 address are removed
and 2 are added (one for IPv4 and one for IPv6), the systemd could think
the service is failing since its frequently restarted. So, reset the
failed before configuration

#### How did you verify/test it?
Verified the tests `test_snmp_fact` and `test_reboot` are PASSED.