NOS-544: test_fib failure related to authentication header (AH) protocol (#654)
### Description of PR
When the randomly selected protocol happens to be authentication header
(AH), packet
is not reaching PTF since it gets dropped earlier in kernel processing
since the
packet is not formed completely. The reason is that the test does not
form a complete packet as required by AH protocol. Need to skip this
protocol (51) similar to the others (41, 47, etc).

Summary:
Fixes # (issue)
https://nexthopai.atlassian.net/browse/NOS-544

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

### Approach
#### What is the motivation for this PR?
test_fib is failing

#### How did you do it?
Skipping the authentication header (AH) protocol from the randomly
selected protocols by hash test.

#### How did you verify/test it?
Test case passes with the fix.

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->