NOS-410: add support to run test_ifa on t0 topologies (#651)
### Description of PR
The destination IP used to send the INT packets was chosen at random,
which worked fine for a T1 topology, where the default route was used to
send the packet on front-panel ports. But on T0 topology the route for
the chosen IP points to downstream host links, causing test failures.

Use one of the upstream peer VM IP address as destination for INT
packets. Use minigraph to find the IP address of peer VM.

Summary:
Fixes # NOS-410

### Type of change
- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
Use minigraph to find the IP address of peer VM, and use that as
destination IP of INT packets.

#### What is the motivation for this PR?
Support T0 topologies

#### How did you verify/test it?
Manual run passes on both t0 and t1 topologies