NOS-3273: [bgp] Wait additionally for the session to come up (#668)
Summary:
Fixes # NOS-3273

### Type of change
- [x] Bug fix

### Back port request
- [x] 202505

### Approach
#### What is the motivation for this PR?
The BGP session doesnt come up within 60 seconds and the test fails due
to that.

#### How did you do it?
Increased the timeout to 90 seconds.

Note: We would have to evaluate FRR why it takes more time to get the
session established, but this seem to happen only when we make the IP
based neighbor to Idle (Admin) and configure interface neighbor. We dont
see this happen if just Interface neighbor is configured.

#### How did you verify/test it?
```
TESTS RUN RESULTS:BEG
+-------------------------------------+-----------------------------------+--------+----------+-------+
| Test                                | Description                       | Result | Duration | Error |
+-------------------------------------+-----------------------------------+--------+----------+-------+
| tests/bgp/test_bgp_link_local_v6.py | skipped 0 errors/failed 0 total 2 | PASS   | 508.93   |       |
+-------------------------------------+-----------------------------------+--------+----------+-------+
TESTS RUN RESULTS:END
```