NOS-3419: Update test_psu.py::test_temperature to handle invalid temperatures (#706)
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit
easier:
-->
### Description of PR
<!--
- Please also include relevant motivation and context. Where should
reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Fix logic issue that would allow test_psu.py::test_temperature to do a
comparison with return value of get_temperature even when it is not
valid

Fixes # NOS-3419

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

### Approach
#### What is the motivation for this PR?
To prevent an error from occurring if a PSU returns a temperature value
of "N/A" due to the PSU being uninstalled.

#### How did you do it?
Skip the rest of the test if the temperature value is invalid.

#### How did you verify/test it?
Ran the test on hum120.
Confirmed it passes nominally.
```
| tests/platform_tests/api/test_psu.py::TestPsuApi::test_temperature | skipped 0 errors/failed 0 total 1 | PASS   | 40.88    |       |
```

Modified the test so that the temperature was "N/A" for PSU 1
```
| tests/platform_tests/api/test_psu.py::TestPsuApi::test_temperature | skipped 0 errors/failed 1 total 1 | FAIL   | 41.65    | Failed: Failed to retrieve temperature of PSU 1 |
```

Modified the test so that the temperature was"invalid_type" for PSU 1
```
| tests/platform_tests/api/test_psu.py::TestPsuApi::test_temperature | skipped 0 errors/failed 1 total 1 | FAIL   | 40.65    | Failed: PSU 1 temperature validation failed: expected float, got: invalid_value (type: str) |
```

Tested on gold107 and gold111
test_psu.py::test_temperature passed
the result summary of `nhtest/nhtests-run` result was FAIL_ENV
This is consistent with behavior of the test before this change.
```
platform_tests/api/test_psu.py::TestPsuApi::test_temperature[dvc134] PASSED [100%]
```

#### Any platform specific information?
Not applicable

#### Supported testbed topology if it's a new test case?
Not applicable: Not a new test case


### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->
Not applicable

Signed-off-by: Caleb Dueck caleb@nexthop.ai