NOS-862: test_mgmtvrf.py should run NTP in mgmt vrf (#724)
The test currently does not attempt to run the NTP service inside of the
mgmt VRF. Since there is a valid yang model for this feature, we should
be testing it. Add a test variant that uses NTP in the mgmt vrf, as well
as the default one.

<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit
easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should
reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Fixes NOS-862

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [x] 202505

### Approach
#### What is the motivation for this PR?

Add test coverage for NTP in mgmt vrf feature.

#### How did you do it?

Parametrize the test_ntp() function on the vrf name.

#### How did you verify/test it?

With the new testcase, the test fails because it hits the issue from
NOS-862. With the fix for that issue in place, the test passes.

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->