NOS-3600: Fix expected destination nexthop list for everflow ECMP test (#735)
### Description of PR
When a new nexthop is added into the ECMP route for the mirror
destination,
test's expected nexthop list is not updated correctly causing the
failure.

Summary:
Fixes # (issue)
https://nexthopai.atlassian.net/browse/NOS-3600

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [ ] Test case improvement


### Back port request
- [x] 202505

### Approach
#### What is the motivation for this PR?
When a new nexthop is added into the ECMP route for the mirror
destination,
test's expected nexthop list is not updated correctly causing the
failure.

#### How did you do it?
Updated the expected nexthop list for the packet to be received on.

#### How did you verify/test it?
Test passes with fix on golden eagle and hummingbird platform

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->