NOS-3633: Bump reboot timeouts in mvrf/test_mgmtvrf.py (#750)
<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit
easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should
reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Re-enable `mvrf/test_mgmtvrf.py`. The community has recently marked it
as skipped for all broadcom asics, but this is a useful test for us so
we should keep running it. It's one of our most basic smoke tests of
mgmt VRF functionality.

Bump the reboot testcase timeouts, as the time for SNMP facts to come up
has increased recently. We should look into why the time has crept up,
but that's a separate issue. For now, this is enough of a fix to keep
the test valid.

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [x] 202505

### Approach
#### What is the motivation for this PR?

This test is failing with timeout issues. We want to keep the coverage
of the actual feature under test, rather than testing reboot times
(which we should also attend to, but no need to lose test coverage in
the meantime).

#### How did you do it?

Increase the timeouts.

#### How did you verify/test it?

Before this change, the test fails all three reboot cases due to timeout
issues.

After this change, the reboot cases all either pass or fail with
different, more relevant issues.

#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->